[#84] TS - Small items to complete
5.1 New/Edit Tag List
5.1.1 Add colour picker to screen.

Add a new field to the Tag List Modal that allows the user to select a
colour and write this value to the database field tag.tagcolor. Example
jquery component here: http://www.eyecon.ro/colorpicker/

5.1.2 Add Entity Selector to screen.

Add a new field to the Tag List Modal that allows the user to select one
or more entities related to the tag. You can use the same type of
control current available for the Class Unit selection.

5.2.1.3 Lesson List (item 3.1.1 in attached email)